accept multiple -f options, and thus parent containers
Add support for multiple -f options, keep a list of all files other than the
first and import them into the container of the first, assuming a lineage.

We unfortunately can't easily follow the lineage ourselves as the parent
tends to be recorded with only the basename, which would necessitate
search-path functionality matching that of krtld to function correctly.
So we cheat.